send_mail_gmail/requirements.txt
# python3.11

python-dotenv